outline how I want stuff reorganised
