--- common/CommonPCH/PCH.hpp
@@ -1,13 +1,0 @@
-// Copyright (C) 2018-2022 - DevSH Graphics Programming Sp. z O.O.
-// This file is part of the "Nabla Engine".
-// For conditions of distribution and use, see copyright notice in nabla.h
-#ifndef _EXAMPLES_COMMON_PCH_HPP_
-#define _EXAMPLES_COMMON_PCH_HPP_
-
-#include <nabla.h>
-
-#include <InputSystem.hpp>
-#include <CCamera.hpp>
-#include <SimpleWindowedApplication.hpp>
-
-#endif // _EXAMPLES_COMMON_PCH_HPP_
