+++ common/include/nbl/examples/PCH.hpp
@@ -1,0 +1,22 @@
+// Copyright (C) 2018-2025 - DevSH Graphics Programming Sp. z O.O.
+// This file is part of the "Nabla Engine".
+// For conditions of distribution and use, see copyright notice in nabla.h
+#ifndef _NBL_EXAMPLES_PCH_HPP_
+#define _NBL_EXAMPLES_PCH_HPP_
+
+
+#include <nabla.h>
+
+// #include "nbl/ui/CGraphicalApplicationAndroid.h"
+// #include "nbl/ui/CWindowManagerAndroid.h"
+
+#include "nbl/examples/common/SimpleWindowedApplication.hpp"
+#include "nbl/examples/common/InputSystem.hpp"
+#include "nbl/examples/common/CEventCallback.hpp"
+
+#include "nbl/examples/cameras/CCamera.hpp"
+
+#include "nbl/examples/geometry/CGeometryCreatorScene.hpp"
+
+
+#endif // _NBL_EXAMPLES_COMMON_PCH_HPP_
