+++ common/include/nbl/examples/common/SBasicViewParameters.hlsl
