--- common/src/camera/CMakeLists.txt
@@ -1,7 +1,0 @@
-# header only currently
-
-#set(NBL_LIB_SOURCES
-#    "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp"
-#)
-
-#nbl_create_ext_library_project(Camera "" "${NBL_LIB_SOURCES}" "" "" "")
