--- common/src/geometry/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(creator EXCLUDE_FROM_ALL)
