+++ common/src/nbl/examples/cameras/CMakeLists.txt
@@ -1,0 +1,7 @@
+# header only currently
+
+#set(NBL_EXAMPLES_CAMERA_LIB_SOURCES
+#    "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp"
+#)
+
+#nbl_create_ext_library_project(ExampleCameras "" "${NBL_EXAMPLES_CAMERA_LIB_SOURCES}" "" "" "")
