+++ common/src/nbl/examples/geometry/shaders/gc.basic.fragment.hlsl
