+++ common/src/nbl/examples/geometry/shaders/gc.cone.vertex.hlsl
