+++ common/src/nbl/examples/geometry/shaders/gc.ico.vertex.hlsl
