+++ common/src/nbl/examples/geometry/shaders/template/gc.basic.vertex.input.hlsl
