+++ common/src/nbl/examples/geometry/shaders/template/gc.cone.vertex.input.hlsl
