+++ common/src/nbl/examples/geometry/shaders/template/gc.ico.vertex.input.hlsl
