+++ common/src/nbl/examples/geometry/shaders/template/gc.vertex.hlsl
