+++ common/src/nbl/examples/geometry/shaders/template/grid.common.hlsl
@@ -31,7 +31,7 @@
     }
 #endif // __HLSL_VERSION
 
-#include "SBasicViewParameters.hlsl"
+#include "common/SBasicViewParameters.hlsl"
 
 #endif // _THIS_EXAMPLE_GRID_COMMON_HLSL_
